Move llm functions to sublib
