+++ tracecat/llm/__init__.py
@@ -1,0 +1,19 @@
+from .ollama import (
+    OllamaModel,
+    async_ollama_call,
+    list_local_model_names,
+    list_local_models,
+)
+from .openai import (
+    OpenAIModel,
+    async_openai_call,
+)
+
+__all__ = [
+    "OllamaModel",
+    "OpenAIModel",
+    "async_ollama_call",
+    "async_openai_call",
+    "list_local_model_names",
+    "list_local_models",
+]
